Update README.md
Hi there! 

Im afraid this software was released under the MIT licence with requires under its condition that the licence be kept intact with this software. Please revert the changes you have made to the license here. We are of course happy for you to use our software and the MIT is one of the most flexible licenses out there, but you must respect the requirements of that license. 

Thank you!